Room Crashing fixed
Fixed an issue where when there were mulutple rooms the site could
crash because it couldnt properly detect new rooms.

Also fixed an issue where when a new room would be created it would
cause redrawing on deleted lines in other rooms lol.

A new issue: when a new user joins someones room it will get the
default ('consulate') select value from the client and set the
current state of the room to that map!